fix: change response diamond to float

+++ Assets/Scripts/ChangeClass/API/CharacterResponseData.cs
@@ -10,7 +10,7 @@ namespace CryptoQuest.ChangeClass.API
         public bool success;
         public string message;
         public int gold;
-        public int diamond;
+        public float diamond;
         public int soul;
         public long time;
         public CharacterData data;
